🏗️ Architect: Modularized lunar skyfield searches
Refactored the 496-line `apts/skyfield_searches/lunar.py` 'God Object' into a modular package `apts/skyfield_searches/lunar/`.

Changes:
- Split orbital logic (supermoons, apogee/perigee) into `orbits.py`.
- Split lunar occultation logic into `occultations.py`.
- Split lunar features and libration logic into `features.py`.
- Maintained the public API via re-exports in `__init__.py`.
- Updated relative imports to support the new package structure.

This improvement follows the Single Responsibility Principle and enhances maintainability of the lunar search algorithms.
